# services/shf-agent-fabric/registry/contracts/openapi.latest.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "SHF Agent Fabric",
        "version": "0.1.0"
    },
    "paths": {
...
        "/admin/align/apps": {
            "get": {
                "tags": [
                    "alignment-admin"
                ],
                "summary": "Apps",
                "operationId": "apps_admin_align_apps_get",
                "parameters": [
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/align/apps/{app_id}/state": {
            "patch": {
                "tags": [
                    "alignment-admin"
                ],
                "summary": "Set State",
                "operationId": "set_state_admin_align_apps__app_id__state_patch",
                "parameters": [
                    {
                        "name": "app_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "App Id"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchState"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/align/containment": {
            "get": {
                "tags": [
                    "alignment-admin"
                ],
                "summary": "Get Containment",
                "operationId": "get_containment_admin_align_containment_get",
                "parameters": [
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "alignment-admin"
                ],
                "summary": "Patch Containment",
                "operationId": "patch_containment_admin_align_containment_patch",
                "parameters": [
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchContainment"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/align/containment/apps/{app_id}/{state}": {
            "post": {
                "tags": [
                    "alignment-admin"
                ],
                "summary": "Force App",
                "operationId": "force_app_admin_align_containment_apps__app_id___state__post",
                "parameters": [
                    {
                        "name": "app_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "App Id"
                        }
                    },
                    {
                        "name": "state",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "enum": [
                                "OFF",
                                "LIMITED",
                                "ON"
                            ],
                            "type": "string",
                            "title": "State"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/align/containment/apps/{app_id}": {
            "delete": {
                "tags": [
                    "alignment-admin"
                ],
                "summary": "Clear Forced",
                "operationId": "clear_forced_admin_align_containment_apps__app_id__delete",
                "parameters": [
                    {
                        "name": "app_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "App Id"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/admin/agents": {
            "get": {
                "tags": [
                    "admin-agents"
                ],
                "summary": "Admin List Agents",
                "operationId": "admin_list_agents_admin_agents_get",
                "parameters": [
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admin/agents/{name}/enabled": {
            "get": {
                "tags": [
                    "admin-agents"
                ],
                "summary": "Admin Get Agent Enabled",
                "operationId": "admin_get_agent_enabled_admin_agents__name__enabled_get",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Name"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "admin-agents"
                ],
                "summary": "Admin Set Agent Enabled",
                "operationId": "admin_set_agent_enabled_admin_agents__name__enabled_post",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Name"
                        }
                    },
                    {
                        "name": "x-admin-key",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "X-Admin-Key"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/EnabledBody"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "EnabledBody": {
                "properties": {
                    "enabled": {
                        "type": "boolean",
                        "title": "Enabled"
                    }
                },
                "type": "object",
                "required": [
                    "enabled"
                ],
                "title": "EnabledBody"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PatchContainment": {
                "properties": {
                    "block_agents": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Block Agents"
                    },
                    "block_capabilities": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Block Capabilities"
                    }
                },
                "type": "object",
                "title": "PatchContainment"
            },
            "PatchState": {
                "properties": {
                    "state": {
                        "type": "string",
                        "enum": [
                            "OFF",
                            "LIMITED",
                            "ON"
                        ],
                        "title": "State"
                    }
                },
                "type": "object",
                "required": [
                    "state"
                ],
                "title": "PatchState"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}